+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes 
 
 - [#1142](https://github.com/cosmos/iavl/pull/1142) Fix race between updating fast node cache and db commit in `NodeDB`.
+- [#1153](https://github.com/cosmos/iavl/pull/1153) Improve `Importer` validation and validate bounds in `CompressImporter` and `deltaDecoder`.
 
 ## 1.3.5, December 27, 2024